==> 1.3_Test_Set_analysis.ipynb <==
# Test set analysis

from import_man import *

df = pd.read_csv("dataset/test.csv")

df.shape

df.head().T

### How many days are the stores open?

df.loc[(df['IsOpen'] == 1)]

### Closed days percentage

((45689-37586)/45689)*100
