==> 3.0_Prepr_train0_date.ipynb <==
# Data preprocessing: Date extraction

**NOTEBOOK GOAL**: Extract attributes: Day, Month and Year from the Date attribute

**DATASET TRANSFORMATION**: `(train|test).csv` >> `preprocessed0_date_(train|test).csv`

# reoder variables to work on train or test dataset
work_on = 'test'
#work_on = 'train'

from import_man import *

from datetime import datetime

df = pd.read_csv('./dataset/' + work_on + '.csv')
df.head()

df['D_Day'] =  df['Date'].apply(lambda x: datetime.strptime(x, '%d/%m/%Y').day)
df['D_DayOfYear'] =  df['Date'].apply(lambda x: datetime.strptime(x, '%d/%m/%Y').timetuple().tm_yday)

df['D_Month'] =  df['Date'].apply(lambda x: datetime.strptime(x, '%d/%m/%Y').month)

df['D_Year'] =  df['Date'].apply(lambda x: datetime.strptime(x, '%d/%m/%Y').year)

df['D_DayOfweek'] =  df['Date'].apply(lambda x: datetime.strptime(x, '%d/%m/%Y').weekday())

df.head(10)

print("ATTRIBUTES LIST:")
pprint(list(df.columns))

print("\n\nSORTED ATTRIBUTES LIST:")
pprint(sorted(list(df.columns)))

### Write to file

# save the dataset
df.to_csv('./dataset/preprocessed0_date_' + work_on + '.csv', index=False)
