==> 3.2_Prepr_train2_OHE.ipynb <==
# Data preprocessing: OHE - One Hot Enconding

**NOTEBOOK GOAL**: Encode categorical variables

**DATASET TRANSFORMATION**: `preprocessed1_imputation_(train|test).csv` >> `preprocessed2_OHE_(train|test).csv`

One Hot Encoding is needed on the categorical variables to use them for regression or any other kind of computation.
Special attention must be paid to Events, since it has many possible values separated by a dash

#### Attirbutes to be encoded

- StoreType
- AssortmentType
- Events

**INDEX**
- [Unique values of attributes to encode](#Unique-values-of-attributes-to-encode)
- [OHE for StoreType](#OHE-for-StoreType)
- [OHE for AssortmentType](#OHE-for-AssortmentType)
- [OHE for Events](#OHE-for-Events)
- [Final check](#Final-check)

# reoder variables to work on train or test dataset
work_on = 'test'
#work_on = 'train'

from import_man import *

df = pd.read_csv('./dataset/preprocessed1_imputation_' + work_on + '.csv')

### Unique values of attributes to encode

print("Unique values of StoreType")
pprint(df.StoreType.unique())

print("\nUnique values of AssortmentType")
pprint(df.AssortmentType.unique())

print("\nUnique values of Events")
pprint(df.Events.unique())

df.head()

### OHE for StoreType

def OHE_StoreType_SuperMarket(value):
    return 1 if value=='Super Market' else 0

def OHE_StoreType_HyperMarket(value):
    return 1 if value=='Hyper Market' else 0

def OHE_StoreType_StandardMarket(value):
    return 1 if value=='Standard Market' else 0

def OHE_StoreType_ShoppingCenter(value):
    return 1 if value=='Shopping Center' else 0

df['StoreType_SuperMarket'] = df.StoreType.apply(OHE_StoreType_SuperMarket)
df['StoreType_HyperMarket'] = df.StoreType.apply(OHE_StoreType_HyperMarket)
df['StoreType_StandardMarket'] = df.StoreType.apply(OHE_StoreType_StandardMarket)
df['StoreType_ShoppingCenter'] = df.StoreType.apply(OHE_StoreType_ShoppingCenter)

### OHE for AssortmentType

def OHE_AssortmentType_General(value):
    return 1 if value=='General' else 0

def OHE_AssortmentType_WithNFDepartment(value):
    return 1 if value=='With Non-Food Department' else 0

def OHE_AssortmentType_WithFishDepartment(value):
    return 1 if value=='With Fish Department' else 0


df['AssortmentType_General'] = df.AssortmentType.apply(OHE_AssortmentType_General)
df['AssortmentType_WithNFDept'] = df.AssortmentType.apply(OHE_AssortmentType_WithNFDepartment)
df['AssortmentType_WithFishDept'] = df.AssortmentType.apply(OHE_AssortmentType_WithFishDepartment)

df.head()

Notice that for instance AssortmentType_WithFishDept has only a few cases (since the mean is 0.004) and it refers to all examples, since every store has more than 700 tuples (one per each observation day), it means that only a few stores (e.g. 2 or 3 stores) have a AssortmentType_WithFishDept, we could discard cases like that or we could treat those stores separately

### OHE for Events

# Possible single values
# 'Fog', 'Hail', 'None', 'Rain', 'Snow', 'Thunderstorm'

events = [
    'Fog',
    'Hail',
    'Rain',
    'Snow',
    'Thunderstorm'
]

df.columns

for event in events:
    df["Events_" + event] = df.Events.apply(lambda x: 1 if event in x else 0)

## Final check

Let's check which are the attributes contained in the dataset

print("\n\nSORTED ATTRIBUTES LIST:")
pprint(sorted(list(df.columns)))

df.head()

### Write to file

# save the encoded dataset
df.to_csv('./dataset/preprocessed2_OHE_' + work_on + '.csv', index=False)
