==> 3.3_Prepr_train_features_extraction.ipynb <==
# Data preprocessing: Feature Engineering

**NOTEBOOK GOAL**: Extract valuable features that can counts for models training

**DATASET TRANSFORMATION**: `preprocessed2_OHE_imputation_(train|test).csv` >> `preprocessed_(train|test).csv`

**INDEX**

- [Deletion of IsOpen = 0 rows](#Deletion-of-IsOpen-=-0-rows)
- [MeanStoreSales feature](#MeanStoreSales-feature)
- [MeanReagionSales feature](#MeanReagionSales feature)
- [Day of the Week considering days adjacency](#Day-of-the-Week-considering-days-adjacency) 

# reoder variables to work on train or test dataset
work_on = 'test'
#work_on = 'train'

import numpy as np 
import pandas as pd

features_to_print = ['StoreID', 'Date']

df = pd.read_csv('dataset/preprocessed2_OHE_train.csv')
if work_on == 'test':
    df_test = pd.read_csv('dataset/preprocessed2_OHE_test.csv')

### Deletion of IsOpen = 0 rows

Let's delete the rows of store closed.

df.head()

#df_test.head()

df = df[df.IsOpen==1]
if work_on == 'test':
    df_test = df_test[df_test.IsOpen==1]

df.head()

#df_test.head(5)

df.drop(['IsOpen'], axis=1, inplace=True)
if work_on == 'test':
    df_test.drop(['IsOpen'], axis=1, inplace=True)

#df_test.head(5)

### MeanStoreSales feature

Add the mean of the sales of the store

df_mean = df.groupby('StoreID')['NumberOfSales'].mean()
def mean_sales(value):
    # since stores ID are from 1000 to 1749 let's subtract 1000
    return df_mean.iloc[value-1000]

df['MeanStoreSales'] = df.StoreID.apply(mean_sales)
if work_on == 'test':
    df_test['MeanStoreSales'] = df_test.StoreID.apply(mean_sales)


#df[10000:10010].head()

print(df_mean.shape)
df_mean.head()

### MeanReagionSales feature

Add the mean of sales for the region (regardless of the store)

df_mean_reg = df.groupby('Region')['NumberOfSales'].mean()

def mean_sales_region(value):
    return df_mean_reg.iloc[value]

df['MeanRegionSales'] = df.Region.apply(mean_sales_region)
if work_on == 'test':
    df_test['MeanRegionSales'] = df_test.Region.apply(mean_sales_region)

features_to_print += ['D_DayOfweek', 'MeanStoreSales', 'MeanRegionSales']

df[df.Date=="01/03/2016"][features_to_print].head().T

#df_test[df_test.Date=="01/03/2018"][features_to_print].head().T

### MeanCustomers feature

Add the mean of the sales of the store

df_mean = df.groupby('StoreID')['NumberOfCustomers'].mean()
def mean_cust(value):
    # since stores ID are from 1000 to 1749 let's subtract 1000
    return df_mean.iloc[value-1000]

df['MeanCustomers'] = df.StoreID.apply(mean_cust)
if work_on == 'test':
    df_test['MeanCustomers'] = df_test.StoreID.apply(mean_cust)

print(df_mean.shape)
df_mean.head()

### MeanCustomers per region feature

Add the mean of sales for the region (regardless of the store)

df_mean_reg = df.groupby('Region')['NumberOfCustomers'].mean()

def mean_cust_region(value):
    return df_mean_reg.iloc[value]

df['MeanRegionCustomers'] = df.Region.apply(mean_cust_region)
if work_on == 'test':
    df_test['MeanRegionCustomers'] = df_test.Region.apply(mean_cust_region)

features_to_print += ['D_DayOfweek', 'MeanStoreSales', 'MeanRegionSales','MeanCustomers', 'MeanRegionCustomers']

df[df.Date=="01/03/2016"][features_to_print].head().T

df_test[df_test.Date=="01/03/2018"][features_to_print].head().T

### Day of the Week considering days adjacency

<https://www.reddit.com/r/MachineLearning/comments/2hzuj5/how_do_i_encode_day_of_the_week_as_a_predictor/>

sorted(df['D_DayOfweek'].unique())

from math import pi, cos, sin

# angle in rad
df['rad_ang'] = (df['D_DayOfweek'] / 7) * (2 * pi)
#df['D_DayOfWeek_cos'] = df.apply(lambda x: cos(x['rad_ang']), axis=1)
#df['D_DayOfWeek_sin'] = df.apply(lambda x: sin(x['rad_ang']), axis=1)
df['D_DayOfWeek_cos'] = df['rad_ang'].apply(lambda x: cos(x))
df['D_DayOfWeek_sin'] = df['rad_ang'].apply(lambda x: sin(x))

features_to_print  += ['rad_ang', 'D_DayOfWeek_cos', 'D_DayOfWeek_sin']
df[features_to_print].head(20).T

if work_on == 'test':
    df_test['rad_ang'] = (df_test['D_DayOfweek'] / 7) * (2 * pi)
    #df['D_DayOfWeek_cos'] = df.apply(lambda x: cos(x['rad_ang']), axis=1)
    #df['D_DayOfWeek_sin'] = df.apply(lambda x: sin(x['rad_ang']), axis=1)
    df_test['D_DayOfWeek_cos'] = df_test['rad_ang'].apply(lambda x: cos(x))
    df_test['D_DayOfWeek_sin'] = df_test['rad_ang'].apply(lambda x: sin(x))

    features_to_print  += ['rad_ang', 'D_DayOfWeek_cos', 'D_DayOfWeek_sin']
    df_test[features_to_print].head(20).T

#features_to_print  += ['rad_ang', 'D_DayOfWeek_cos', 'D_DayOfWeek_sin']
#df_test[features_to_print].head(20).T

# let's drop the angle column since it has the same menaing of the day of week
df.drop('rad_ang', inplace=True, axis=1)
if work_on == 'test':
    df_test.drop('rad_ang', inplace=True, axis=1)
df.head(1).T

df.shape

df_test.shape

## Write to file

df.to_csv('./dataset/preprocessed_train.csv', index=False)
if work_on == 'test':
    df_test.to_csv('./dataset/preprocessed_test.csv', index=False)

## Additional sales of Store per month (tested and was not successful)

break_here_exec

#dfmean = df.groupby(['StoreID','D_Month'])['NumberOfSales'].mean()
dfmean.head().T

dfmean

dfmean = dfmean.reset_index()
dfmean.head().T

store=1003
m=5
dfmean[(dfmean.StoreID ==store) & (dfmean.D_Month ==m)]['NumberOfSales'][(store-1000)*12+m-1]

df['Mean_month_sale']

df.loc[:10,'Mean_month_sale']=dfmean[(dfmean.StoreID ==1000) & (dfmean.D_Month ==1)]['NumberOfSales'][0]

store=1001
month=3
df.loc[(df.StoreID ==store) & (df.D_Month ==month),'Mean_month_sale']=\
dfmean[(dfmean.StoreID ==store) & (dfmean.D_Month ==month)]['NumberOfSales'][(store-1000)*12+month-1]

for store in (1000,1748):
    for month in (1,12):
        df.loc[(df.StoreID ==store) & (df.D_Month ==month),'Mean_month_sale']=\
                dfmean[(dfmean.StoreID ==store) & (dfmean.D_Month ==month)]['NumberOfSales'][(store-1000)*12+month-1]

#def mean_sales_month(store,month):  #takes too much time
#    # since stores ID are from 1000 to 1749 let's subtract 1000
#    return dfmean[(dfmean.StoreID ==store) & (dfmean.D_Month ==month)]
