==> 8.1_RFR-XGB_Ensemble_final.ipynb <==
# RFR-XGB Ensemble
In general we saw that XGB overall is better (when sampling randomly train and test or in January and February tests), however, Random Forest error does not have a high variance and is better when predicting on March and April, thus we can do an average of the predicted number of sales of every Store in every day, leading to an overall result which is slightly better than the average of the 2 errors.

The other possiblity that we will evaluate is to do an average of the two predictions at a higher level, when we have already grouped by StoreID and Month, this could be even better if one of the model for instance tends to overestimate the predicted value and the over one underestimates it.

from import_man import *
from BIP import get_BIP_error

### Loading XGB and RFR

#RFR = pd.read_csv("dataset/test4_53_Model_RFR_on_prep.csv")
#XGB = pd.read_csv("dataset/test4_64_Model_XGBoost_final.csv")

#RFR = pd.read_csv("dataset/test2_53_Model_RFR_on_prep.csv")
#XGB = pd.read_csv("dataset/test2_63_Model_XGB.csv")

#RFR = pd.read_csv("dataset/test3416_53_RFR.csv")
#XGB = pd.read_csv("dataset/test3416_63_XGB.csv")

RFR = pd.read_csv("dataset/testsplit_53_Model_RFR_on_prep.csv")  #it sucks with these
XGB = pd.read_csv("dataset/testsplit_63_Model_XGB.csv")

#### XGB BIP error

eXGB=get_BIP_error(XGB)

#### RFR BIP error

eRFR=get_BIP_error(RFR)

ens = XGB[['NumberOfSales','_NumberOfSales']]

ens['R_NumberOfSales']=RFR['_NumberOfSales']

ens['avg_NumberOfSales']=(ens['_NumberOfSales']+ens['R_NumberOfSales'])/2

ens.head(10)

XGB['_NumberOfSales']=ens['avg_NumberOfSales']

#### Ensemble BIP error

eEns=get_BIP_error(XGB)

XGB.to_csv('./dataset/test4_81_AVG_RFR_XGB.csv', index=False)

#### Average of the initial BIP errors

(eXGB+eRFR)/2

## From here on is not something we can really do, it's just out of curiosity and if there is a pattern

diff=XGB['NumberOfSales']-XGB['_NumberOfSales']

diff.head(15).T

diff.mean()

XGB['_NumberOfSales']=XGB['_NumberOfSales']+diff.mean()

eXGB=get_BIP_error(XGB)
