==> 8.2_Ensemble_average.ipynb <==
# Models ensemble by averaging

In general we saw that XGB overall is better (when sampling randomly train and test or in January and February tests), however, Random Forest error does not have a high variance and is better when predicting on March and April, thus we can do an average of the predicted number of sales of every Store in every day, leading to an overall result which is slightly better than the average of the 2 errors.

The other possiblity that we will evaluate is to do an average of the two predictions at a higher level, when we have already grouped by StoreID and Month, this could be even better if one of the model for instance tends to overestimate the predicted value and the over one underestimates it.

**NOTEBOOK GOAL**: Ensemble of models created by averaging the preditions

Averaged predictions:

- **1_RFR** - Notebook 5.3 Random forrest
- **2_XGB** - Notebook 6.4 XGBoost

# reoder variables to work on train or test dataset
predict_for_submission = True
#predict_for_submission = False

from import_man import *
import collections

from BIP import get_BIP_error, apply_BIP_submission_format

### Load predicted tests

**NOTE** If you cannot load the followig datasets, please go to the corresponding notebook and run it to generate the related dataset file. 

dfs_dict = collections.OrderedDict()
# the following dataset will be evaluated

if predict_for_submission:
    dfs_dict['RFR'] = pd.read_csv('./dataset/submission_53_RFR_on_prep.csv')
    dfs_dict['XGB'] = pd.read_csv('./dataset/submission_64_XGBoost_final.csv')
else:
    months='m12'
    dfs_dict['RFR'] = pd.read_csv('./dataset/test_' + months + '_53_RFR_on_prep.csv')
    dfs_dict['XGB'] = pd.read_csv('./dataset/test_' + months + '_64_Model_XGBoost_final.csv')

list(dfs_dict.values())[0].head().T

# let's apply the apply_BIP_submission_format to all the dataframes
if predict_for_submission:
    for mdl_lbl, df in dfs_dict.items():
        dfs_dict[mdl_lbl] = apply_BIP_submission_format(df,real_submit=True)
else:
    for mdl_lbl, df in dfs_dict.items():
        dfs_dict[mdl_lbl] = apply_BIP_submission_format(df)

We decided to work with the average of the already summed up value because it reduces more the error

# make a copy of the first dataframe in order to use it as data structure
df_ens = list(dfs_dict.values())[0].copy()
df_ens['NumberOfSales'] = 0

stores_to_spot = [1000, 1245, 1300, 1301] 

# Let's spot some random rows in order to prove that the average works correctly
if predict_for_submission:
    for mdl_lbl, df in dfs_dict.items():
        print('................................ ' + mdl_lbl + '................................')
        print(df.loc[df.StoreID.isin(stores_to_spot)][['StoreID', 'NumberOfSales']].head(40))
else:
    for mdl_lbl, df in dfs_dict.items():
        print('................................ ' + mdl_lbl + '................................')
        print(df.loc[df.StoreID.isin(stores_to_spot)][['StoreID', 'Target', 'NumberOfSales']].head(40))

if predict_for_submission:
    print(df_ens[['StoreID', 'NumberOfSales']].head())
else:
    print(df_ens[['StoreID', 'Target', 'NumberOfSales']].head())

# sum up all
for mdl_lbl, df in dfs_dict.items():
    df_ens['NumberOfSales'] += df['NumberOfSales']

# divide by their number
df_ens['NumberOfSales'] /= len(dfs_dict)

if predict_for_submission:
    print(df_ens.loc[df_ens.StoreID.isin(stores_to_spot)][['StoreID', 'NumberOfSales']].head(40))
else:
    print(df_ens.loc[df_ens.StoreID.isin(stores_to_spot)][['StoreID', 'Target', 'NumberOfSales']].head(40))    

## Write to file

if predict_for_submission:
    df_ens.to_csv('./dataset/submission_82_Ensemble_average.csv', index=False)
else:
    df_ens.to_csv('./dataset/test_' + months + '_82_Ensemble_average.csv', index=False)
