==> 9.10_Submission.ipynb <==
# Create submission test

**NOTEBOOK GOAL**: Create the dataset with predictions to be submitted.


Used model for the submission:

> **3_ENS_AVG** - Notebook 8.2 Ensemble by averaging all the previous models

from import_man import *
import collections

from BIP import apply_BIP_submission_format

### Load test dataset with predictions

**NOTE** To load the followig dataset, please go to the corresponding notebook and run it (test mode) to generate the related dataset file. 

# dataset to use to create the one for the submission
df = pd.read_csv('./dataset/submission_82_Ensemble_average.csv')

df.head()

df.shape

### Year check

#years = df.D_Year.unique()

# let's be sure that the year is only one
#assert(len(years) == 1)

# let's be sure that the year is 2018
#assert(years[0] == 2018)

### Months check

months = df.Month.unique()

# let's be sure that the month are exactly 2
assert(len(months) == 2)

# let's be sure that the year is 2018
assert(sorted(months) == [3, 4])

### Apply submission format

df['NumberOfSales'] = pd.to_numeric(df['NumberOfSales'].round(), downcast='integer')

# apply the submission format
#df = apply_BIP_submission_format(df, real_submit=True)

df.head(10)

df.NumberOfSales.sum()

### Write to file

df.to_csv('./dataset/submission_893784.csv', index=False)
